cros-ide: Suggest extensions to install based on language
Add a mechanism to suggest extensions to install based on the language
of the opened files.
If the user agrees, the extension will be automatically installed.
If the user disagrees, the suggestion will not be shown again for the
current session.

As a first usage, suggest Go extension for Golang.

BUG=b:227293635
TEST=manually. https://screenshot.googleplex.com/3frdyQThAae6WFD

Change-Id: Ieef28ca5cf991fa4cb65dcaf5ebc0f19d5551b7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3567842
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>