cros-ide: Fix tast linting errors.
cros lint may fail for file paths with `tast` that are not in
platform/tast. This is because the cwd is incorrectly set to
platform/tast in these cases, creating incorrect file paths for
tast-lint. This change fixes this by specifying the cwd for
platform/tast-tests files and platform/tast-tests-private files.

BUG=b:241348494
TEST=tested using development host
TEST=https://screenshot.googleplex.com/489iyR36KZQCrta

Change-Id: I9e0f6b95641ffb919771ff870f2baad20101768d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3813504
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: Uwem Wilson <uwemwilson@google.com>
Tested-by: Uwem Wilson <uwemwilson@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>