cros-ide: change when we call the spell checker.
Spell checker is similar to cros lint, so we call it during the same
events as in cros_lint.ts. It doesn't make sense to call it whenever the
file is modified, because it reads what on disk.

BUG=b:217287367
TEST=manually (by looking at the log) verified that spell checker
is called when files are opened and saved.

Change-Id: Ic00a7d47bac20ef700f514aba81a86fd92c50e84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3853998
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>