cros-ide: Remove hard-coded mapping in packages/
To retrieve the platform2 package name from a filename, we no longer use
the hard-coded mapping, but compute it at runtime. This patch removes
the hard-coded mapping.

We do this refactoring before moving the packages component to a common
place for the platform2 unit test feature to use it.

BUG=b:254145837
TEST=repo upload

Change-Id: Ib7be7ebe836304143866534a1d54a53c5ae9ff7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3974055
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>