cros-ide: Add an option to release script for patch release
The release script has a mechanism to prevent the user to publish a
revision that is not merged, but before this patch it was just checking
cros/main. It was problmatic for patch release, for which the
revision is contained in the remote release branch, not cros/main. This
patch solves the problem allowing the user to specify the remote branch.

BUG=b:258998028
TEST=Not tested yet. It's difficult to test it because this function asserts HEAD is in FETCH_HEAD, but it is never true until this patch is committed.

Change-Id: I2f0d94caf8a1d0e5e432896c267b6dc797a33e42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4022438
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>