ide: Address warnings on building tests
npm run build-tests shows warnings about these `require`s. We actually
don't have to worry about them, but this change fixes the cause of the
warnings anyway to reduce the visual noise.

These files are compiled to ESM and then transpiled to CJS. The warning
happens on compilation to ESM, because the compiler currently doesn't
support compilation of `require`s into `import`s, and keeps them as is.
It shouldn't have been an issue, because we transpile the intermediate
ESM file into CJS anyway.

BUG=b:275545000
TEST=npm t

Change-Id: I168faa6887fd4508552e7b138b36ddb2e47efc7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4705638
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>