ide: Update the version dev_install.sh creates to be more informative
Before this change ./dev_install.sh was incrementing the patch version
on building the extension from the source code and installing it. It
means VSCode dones't automatically installs a new patch release if
there's the extension built from the source by ./dev_install.sh
because it only supports major.minor.patch for extension versions and
semver pre-release tags are not supported.

With this change, ./dev_install.sh no longer updates the patch version.
In addition to that it makes the version more informative by adding
build metadata such as the commit hash where the extension was built
from to the version.

BUG=b:251715696
TEST=Run ./dev_install.sh and confirm the extension is installed with
the version having the HEAD's commit hash.

Change-Id: I735a401e2e7245b701f334391411f99d9099d32a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4719527
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>