ide: boards and packages: add a flag to develop build and deploy
This change adds a flag to develop build and deploy features under
which.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload

Change-Id: Ib3c2658c0416d86e6ca9f3a2915179c370b4eae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4852788
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>