ide: boards and packages: move an integration test to unit test
Before making another change on boards and packages, move the
integration test for it to unit test, because from the next change we
start registering a command when the class for boards and packages is
instantiated and it produces duplicated registrations issue if we keep
it as an integration test.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload

Change-Id: Ie8d2d0a199ec945925d7981b1805fa523e42080a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4828190
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>