ide: boards and packages: enable the v2 by default
Boards and packages v2 should be now feature parity with the existing
boards and packages view.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload

Change-Id: I16aed28b6560f20440c559e2a82969c936250727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4849111
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>