ide: boards and packages: add a context menu to generate coverage
This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=open codelab_test.cc, right click the codelab package item, click
the menu to generate coverage, and confirm the coverage information is
shown.

Change-Id: I9d7a5e3e46b20a02349b0eb6ae68150e4bcef144
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4828193
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>