ide: boards and packages: add favorite context menu for package
This change allows the user to add packages to favorite. Favorite
packages are listed first to be easy to access.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=https://screenshot.googleplex.com/5uVDZmD595aAFzz.png

Change-Id: Ia4a418eff459f660c13e66d5c4fd338af3b75efb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4848594
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>