ide: boards and packages: add favorite context menu
This change allows the user to add categories to favorite. Favorite
categories are listed first under each board item to be easy to
access.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=https://screenshot.googleplex.com/XGsCBFsihn7W8Ej.png

Change-Id: I66f02d7a2c4c6c4928fb9228928a33e3695c1f30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4845914
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>