ide: boards and packages: rename a command
Rename a command for the current boards and packages so that the name
doesn't conflict later on implementing the v2.

This is a part of the changes to overhaul boards and packages view.
See the commit message of https://crrev.com/c/4791430 for all the TODOs
we will fix through these changes.

Design doc: http://go/chromiumide-dd-build-and-deploy

BUG=b:214336231
TEST=npm run preupload
TEST=click the + button on boards and packages view, enter a package to
workon and confirm the package is worked on.

Change-Id: I7146e711cab18b381486184428112ea090d325da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4835434
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>