ide: Add a wrapper function to register a command
This is part of the changes to achieve the following objective.

vscode.commands.registerCommand ignores errors thrown from an async
function registered as a callback. This behavior is inconvenient for
us to debug command failures. Let's create a wrapper of
registerCommand which handles errors thrown from the callback, and
force its use throughout ChromiumIDE.

BUG=b:298880122
TEST=npm run unit-test

Change-Id: Ib37c7c5244d507f1649c46dc9960fbd8d9882b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4835658
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>