ide: Use tast run command for debugging
This CL changes to use tast run command when debugging instead of run_debugger.py to handle parametalized tests.
This change adds to dispose of the task provider after use. The reason is to check that the executed command changes when we change the selection of parametalized tests to debug.

BUG=b:255632861
TEST=npm run preupload
TEST=Manually run “Debug Test” in the go file (e.g. src/platform/tast-tests/src/go.chromium.org/tast-tests/cros/local/bundles/cros/example/param.go)

Change-Id: I9f46ce11421c9a0f4fa4491c9398594d39400513
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4835649
Tested-by: Chiaki Ushimaru <uchiaki@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Chiaki Ushimaru <uchiaki@google.com>