ide: Update use of deprecated FakeExec.on method in fakes/cros_sdk (1)
This change is a part of changes to update the use of deprecated
FakeExec.on method in fakes/cros_sdk. In this change we deprecate the
existing function and add a new function that doesn't use FakeExec.on.
We update a test to use the new function to demonstrate how to use it.

BUG=b:299188409
TEST=npm run preupload

Change-Id: I7671b67943ce711f597a21d63cf18e1c9bea9ddd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4844895
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>