ide: gerrit: draft editing (3): support deleting draft
This change adds a button to draft changes to discard the draft. This
feature is not under a flag given it completes within this CL and will
be useful in itself.

This is part of the changes to support editing draft comments on
ChromiumIDE Gerrit support.

BUG=b:260946877
TEST=npm run preupload

Change-Id: I5d2a9f777f458e7e32272129f7a5dd7e36cb86b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4858659
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>