ide: portage parser: get position of char with vscode.TextDocument API
The source of truth should be coming from vscode instead of implementing
the mapping function on our own. This is necessary because a \t
character could be expanded to different width depending on file types
and user setting, and caused wrong offsets in some cases when it was
hardcoded to 2. Updated fake TextDocument implementation (\t has width
2).

BUG=b:303398643
TEST=npm run preupload
TEST=before change: https://screenshot.googleplex.com/4Kvvu4kmWf6aaaL
TEST=after change: https://screenshot.googleplex.com/AyCVawTgu5z3AwJ

Change-Id: I9ac30d3313eac0355435a3163ef80c1f617b7ddf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/4970336
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>