ide: device_management: refactor DeviceClient
sshIdentity and logger is independent of the queried device (hostname)
and should be the same in one vscode session.

This is to prepare for next optimization to regularly refresh device
information in the background (via device repository) by reading its lsb
release, so that we can use the cached value immediately for flashing
image and deploying package to device.

BUG=b:308059094
TEST=npm t

Change-Id: I5ac06c4d3e9c2234325b2fdb0cfd122ed04c478b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5098784
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>