ide: decrease Gerrit poll frequency
Poll gerrit every 1 minute instead of 10 seconds.
In case there are many changes, polling Gerrit server every 10
seconds for each changes can be too much and could cause 429 error.

The downside of this change is that comments by the reviewer may not
appear on the editor in 1 minute maximum, but it should be acceptable.

BUG=b:216048068
TEST=npm t

Change-Id: I269df8fdab5d465a6a073bd61c8e06d9de1a0b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5144031
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>