ide: https.getOrThrow throws error on status code 404 (not found)
A GET request should not be expecting 404 not found error in general.
This allows Https.getOrThrow to always return string (never undefined).

The usage in gerrit is special since user is expected to have local
change not yet uploaded to gerrit as part of their workflow. This is now
handled (catch 404 error thrown by Https.getOrThrow and resolve as
undefined) from the gerrit side.

BUG=b:317141023
TEST=npm t

Change-Id: I14bd4ebd3d9f50dc94c46f7650360a52ea11885d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5199120
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>