BoardOrHost refactor to Board | Host
Separate Board and Host as two different classes implementing the same
CommonInterface, and BoardOrHost is a type being union of the two.

This allows usages that want only Board (or only Host, but none so far)
to request the exact type and be benefited from the type check instead
of checking at runtime and simplifies error handling.

Most existing callsites could be kept the same or very similar (changing
only how the class or functions are imported).

BUG=b:322415636
TEST=npm t
TEST=manual: open IDE and check board and packages features etc work

Change-Id: Ic768fdd317393f18eef732c9452b882b3e2ddd81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5237788
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>