Restrict valid string values as a CrOS image type
Primarily to make sure the compatibility check is correct and not given
an unexpected image type, and when searchign for available prebuilt
images on gs://chromeos-image-archive.

The check is not enforced in DeviceClient since the image type on device
does not matter for the other features to be used.

Next step: support public image stored in gs://chromiumos-image-archive,
which has USE=-cros-debug.

BUG=None
TEST=npm t

Change-Id: I22bd054b3e80d9cb240d6b4dd42c6e4815295515
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5316542
Auto-Submit: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>