cros format: do not format files matching nearest .presubmitignore
We will be expanding scope of cros format to all files.
For users with format on save (and similar features) enabled, this might
be too aggressive and format files that are explicitly exempted.
If the file matches a pattern in .presubmitignore of any of its ancestor
directories, do not format it.

Implementions are based on the functions in the pre-upload script:
https://source.corp.google.com/h/chromium/chromiumos/codesearch/+/main:src/repohooks/pre-upload.py?q=_get_ignore_wildcards

BUG=b:329601336
TEST=npm t
TEST=manual: add error to a file, see that it is reverted on formatting
TEST=manual: add the filename (and as a matching pattern) to the \
  .presubmitignore in its repo; repeat and see that it is not formatted

Change-Id: I6fcc80bd3cbc03b1894c9b6d332a825b272be570
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5439516
Commit-Queue: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>