depot_tools: set path if not found
If `which cros` can't be resolved, it's almost certain that the user
doesn't have a version of depot_tools installed in their home directory
or the custom path. Prompt the user for a directory in this case and set
the configuration variable for the extension. It will not validate
immediately but will prompt again if depot_tools still cannot be found.

BUG=b:322278198
TEST=npm t
TEST=moved depot_tools, was prompted, checked settings were updated
TEST=works normally and doesn't prompt when depot_tools is specified

Change-Id: Ia26bf007e06c553c7d5f9bc7a7c4775aca47b779
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5417676
Reviewed-by: Keigo Oka <oka@chromium.org>
Tested-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>