cpp_xrefs: Register handlers only when needed
This change doesn't introduce behavior changes because currently this
class is instantiated only when registering compdb generator to the
class instance is needed. However in a later change we will decouple the
instantiation and the registerations, and this change is needed to make
the change as well transparent to the user.

BUG=b:333319601
TEST=npm t

Change-Id: I9608e7a2ea0a14d1966653c89e973dc080f00664
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5546765
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>