tast: Detect symlink and provide buttons to address the issue
Xrefs don't work for Go files under symlink `cros`. This change lets
chromiumide detect it if such files are opened and warn the user. The
warning will have a button to open the destination of the symlink for
xrefs to work, and optionally update the workspace config to exclude the
symlink from the files VSCode shows.

BUG=b:342932056
FIXED=b:342932056
TEST=https://screencast.googleplex.com/cast/NTI3MDE2ODE5NTMwMTM3NnwxZGMxZTJkZS01Yg
TEST=npm t

Change-Id: I753432055b21d0757b9e19f8e0c8d72abe351e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5573205
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>