ebuild: Implement hover on LSP
Implement the same functionality as PortageReferenceHoverProvider
provides in LSP and add the equivalent test cases.

The logic for hover impl is now in two places. We will remove the
duplication soon after migration completes.

Metrics are curretly not sent on the LSP impl. It will be added
separetely.

BUG=b:343112937
TEST=npm t

Change-Id: Ib940ca8f4533c44c312c0f227c061ae445e258a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5584425
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>