gcert: Read error message from syslog and append it to message
Just append the syslog message to the error message. We could assume the
message contains a JSON and parse it to provide a prettier error
message, but it increases a dependency to the particular message
format and can be a fragile construct, so I didn't choose the direction.

BUG=b:344489079
FIXED=b:344489079
TEST=npm t
TEST=https://screenshot.googleplex.com/jDEsSzB5HSztyu2

Change-Id: I86d16339bdf35f8baf2ff57b4678959a0e308a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5599868
Reviewed-by: Grace Cham <hscham@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>