lint: Stop resolving fsPath to realpath
This change removes the logic that looks unnecessary.

This logic was added in https://crrev.com/c/3651925 but it and the
associated bug lack explanation about why running realpath for
document.uri.fsPath is necessary. Assuming the file being linted is in
chromeos tree, resolving it to the realpath should not be necessary.

BUG=b:337138396
TEST=npm t

Change-Id: I9914b187e864eb125ad2f1d97a23db0c32fc6c27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5623871
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>