default board: increase status bar item discoverability when not set
The default board status item was hidden before user has set a value to
it which makes it less discoverable and less likely for them to set it.
The item will now always be shown and in such case with makes it
explicitly known to user that they haven't set it with a warning
background color.

Also add a tooltip to let users know that they can update the config by
clicking the item.

BUG=b:337134451
FIXED=b:337134451
TEST=https://screenshot.googleplex.com/5hEGZbezRLtL6J4
TEST=status reset to normal background when a board is chosen

Change-Id: I27233e276dbdee4ce654f4456dc7b67a943f7e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5624774
Tested-by: Grace Cham <hscham@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>