cros format: save the file before running `cros format` on it
So that `cros format` is applied on the latest state of document in the
editor.

The previous fix (running on a copy in tmp/) may produce a different
result since the formatter's behavior depends on the content of
nearby directories (e.g. existence of .clang-format for formatting cpp
files).

BUG=b:349692010, b:343619614
TEST=src/platform2/codelab.cc formatted correctly
TEST=npm t

Change-Id: I14758e1532df3af3fede1a6fdba4b4d07dd7fa00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5662596
Reviewed-by: Keigo Oka <oka@chromium.org>
Auto-Submit: Grace Cham <hscham@chromium.org>
Tested-by: Grace Cham <hscham@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>