gtest: Update the logic to detect the executable including the test
Before this change, there were assupmtions that 1. gtest executables
always contain the exact file path to gtest.h or libgtest.so, and 2. the
executables contain the exact file path to testrunner.cc, and we used
(1) for filtering gtest executables and (2) for getting the directory to
platform2 for path substitutions for debugging. However, in actuality
the both assumptions were incorrect as we tested it on
memory_info_test.cc. (*)

In this change, we mitigate the issue by refining the heuristics to use
partial match rather than the exact match for both (1) and (2).
Moreover, we add an executable as a candidate even if the step (2)
fails because now that we can't rule out the possibility of a gtest
executable containing no platform2 paths.

This change is just a mitigation, which hasn't been proven to work
always. We should revisit the algorithm to ensure its robustness.

(*) The expected string and actual:

- want libgtest.so, got libgtest.so.1.13.0
- want usr/include/gtest/gtest.h, got
  /build/brya/usr/include/gtest/gtest.h
- want .../platform2/common-mk/testrunner.cc, got
  .../platform2/common-mk

BUG=b:356025861
TEST=open memory_info_test.cc, click a button to run a test, confirm the
test passes.
TEST=same for codelab_test.cc.

Change-Id: Ia2942742451931a2aa7572df87a20a0fe5133646
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/5740961
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Grace Cham <hscham@chromium.org>