ide: java/ls: Avoid scanning the whole source checkout
This patch introduces a field to InitializationOptions to specify the
initial source directories to scan. Previously, the server scanned the
whole Chromium source tree, which took a significant amount of time. By
limiting targets to Java source directories only, the initial scan
finishes in a second.

We don't have a plan to upstream this patch.

BUG=b:383450257
TEST=mvn test
TEST=npm run build

Change-Id: I3c9a6521162aa6ebc79ecef894442b85c5ebc4a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6090782
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>