ide: java/ls: Autoimport completed class names
This patch introduces a new interface to represent the import order. We
also provide its two implementations: the simple one (all imports in a
single section with lexicographical order), and the chromium one
(imports are organized into sections as defined in the style guide).

The interface is used in two places: the existing logic to add an import
as a quick fix, and the new logic to add an import on completing class
names.

The style can be configured by user settings. For ChromiumIDE, we fix it
to the chromium style.

BUG=b:383450257
TEST=mvn test
TEST=npm run build

Change-Id: Id47dd51746eb6c32dc198f1e47ebf43fa26f4a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6097043
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>