ide: java/ds: Improve variables view
We used to have two scopes of variables, "locals" and "arguments".
However, VM somehow returns incorrect results for
LocalVariable#isArgument(), so we presented confusing views to users.
Since it's not very important to distinguish arguments and local
variables, we now include them in the same "locals" scope.

Instead, we introduce the new "fields" scope containing fields of the
current object (aka "this").

And finally, we now support inspecting the hierarchy of objects.

BUG=b:385660974
TEST=mvn test -Dtest=JavaDebugServerTest
TEST=install a breakpoint and see variables list
TEST=http://screen/5w8sCBZtusoUMML

Change-Id: Iae5d318f0256f792ab581a69618408b79d3a0f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6193072
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>