ide: java/ls: Show the entire javadoc on hovering/completion
We used to show only the first sentence on hovering/completing
identifiers. This patch allows parsing the whole javadoc.

MarkdownHelper used to manually parse javadoc and thus was hard to
extend to support various javadoc syntaxes. So this patch overhauls it
to use DocTreeScanner to scan the parsed javadoc AST.

BUG=b:385660320
TEST=mvn test -Dtest=MarkdownHelperTest

Change-Id: I72c3375c5f999500304b3608635855273ebdd355
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/infra/ide/+/6217740
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>